+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Removed Sign-up Fee feature (gh-329)
 * Removed “Payment Methods” option in subscription level setup page (gh-330)
 * Removed “Level Groups” feature. Use subscription level relations instead. (gh-331)
+* Removed “States” feature. This feature is no longer used since we removed tax calculations. (gh-333)
 
 # 6.2.3